+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- exiting comparison if no new IB transactions found
+
 ## [0.4.3] - 2023-01-11
 
 ### Fixed